// src/app/core/api/model/loginResponseDto.ts
/**
 * AGORA API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserSummaryDto } from './userSummaryDto';

export interface LoginResponseDto {
  accessToken?: string;
  tokenType?: string;
  expiresIn?: number;
  user?: UserSummaryDto;
}