// src/app/core/api/model/registerResponseDto.ts
/**
 * AGORA API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface RegisterResponseDto {
  id?: string;
  email?: string;
  firstName?: string;
  lastName?: string;
  accountType?: RegisterResponseDto.AccountTypeEnum;
  accountStatus?: RegisterResponseDto.AccountStatusEnum;
}
export namespace RegisterResponseDto {
  export const AccountTypeEnum = {
    Autonomous: 'AUTONOMOUS',
    Tutored: 'TUTORED',
  } as const;
  export type AccountTypeEnum = (typeof AccountTypeEnum)[keyof typeof AccountTypeEnum];
  export const AccountStatusEnum = {
    Active: 'ACTIVE',
    Deleted: 'DELETED',
  } as const;
  export type AccountStatusEnum = (typeof AccountStatusEnum)[keyof typeof AccountStatusEnum];
}